Add files via upload
See the dashboard that I created based on UK accident data. In this project I used several Pivot tables to analyse the data and concise it using calculated items.
In order to see the full project please download the excel file and also enable editing. Without this the data is not able to feed into the dashboard.